+++ src/ElemCo.jl
@@ -8,6 +8,7 @@ module ElemCo
 include("version.jl")
 include("../lib/TREXIO/src/TREXIO.jl")  # Include standalone TREXIO module
 include("infos/abstractEC.jl")
+include("tools/mtensoroperations.jl")
 include("tools/descdict.jl")
 include("tools/outputs.jl")
 include("tools/utils.jl")